Exception Handling

+++ rest_api/src/main/java/com/example/rest_api/restcontroller/EmployeeController.java
@@ -2,8 +2,11 @@ package com.example.rest_api.restcontroller;
 
 import com.example.rest_api.entity.Employee;
 import com.example.rest_api.exceptions.EmployeeExceptions;
+import com.example.rest_api.exceptions.EmployeeNotFoundException;
 import com.example.rest_api.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -17,16 +20,30 @@ public class EmployeeController {
     @Autowired
     private EmployeeService employeeService;
 
-    @GetMapping()
+    /*@GetMapping()
     public List<Employee> findAll() {
         return employeeService.getEmployeeList();
+    }*/
+
+    @GetMapping()
+    public ResponseEntity<List<Employee>> findAll() {
+        return new ResponseEntity<>(employeeService.getEmployeeList(), HttpStatus.OK);
     }
 
-    @GetMapping("/{id}")
+    /*@GetMapping("/{id}")
     public Employee findEmployee(@PathVariable Long id) {
         Employee emp = employeeService.getEmployee(id);
         if (emp != null) return emp;
-        else throw new EmployeeExceptions(id);
+        else throw new EmployeeNotFoundException(id);
+    }*/
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Employee> findEmployee(@PathVariable Long id) {
+        Optional<Employee> emp = employeeService.getEmployee(id);
+        if (!emp.isPresent()) {
+            throw new EmployeeNotFoundException(id);
+        }
+        return new ResponseEntity<>(emp.get(), HttpStatus.OK);
     }
 
     @PostMapping()
@@ -42,7 +59,10 @@ public class EmployeeController {
 
     @DeleteMapping("/{id}")
     public HashMap<String, String> deleteEmployee(@PathVariable Long id) {
-        this.findEmployee(id);
+        Optional<Employee> emp = employeeService.getEmployee(id);
+        if (!emp.isPresent()) {
+            throw new EmployeeNotFoundException(id);
+        }
         employeeService.deleteEmployee(id);
         HashMap<String, String> map = new HashMap<>();
         map.put("message", "Delete successfully!");
